Moved steps specific for CF to a sub dir
