+++ src/Steps/CloudFoundry/StepAuth.php
@@ -1,16 +1,16 @@
 <?php
 /**
- * Step to log off a Cloud Foundry instance.
+ * Step to log in to a Cloud Foundry instance.
  */
 
-namespace Graviton\Deployment\Steps;
+namespace Graviton\Deployment\Steps\CloudFoundry;
 
 /**
  * @author   List of contributors <https://github.com/libgraviton/graviton/graphs/contributors>
  * @license  http://opensource.org/licenses/gpl-license.php GNU Public License
  * @link     http://swisscom.ch
  */
-final class StepLogout extends AbstractStep
+final class StepAuth extends AbstractStep
 {
 
     /**
@@ -22,7 +22,9 @@ final class StepLogout extends AbstractStep
     {
         return array(
             $this->cfCommand(),
-            'logout'
+            'auth',
+            $_SERVER['SYMFONY__DEPLOYMENT__CF_LOGIN_USERNAME'],
+            $_SERVER['SYMFONY__DEPLOYMENT__CF_LOGIN_PASSWORD'],
         );
     }
 }
